Extract <ExerciseStatusDot /> into a separate component
